requirements.txt
apex==0.9.10dev
datasets==2.2.2
filelock==3.7.1
fire==0.4.0
gpytorch==1.7.0
matplotlib==3.5.2
netcal==1.2.1
numpy==1.22.4
opacus==1.1.2
packaging==21.3
pandas==1.4.2
Pillow==9.2.0
prv_accountant==0.1.1.post1
pytest==7.1.2
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentence_transformers==2.2.0
setuptools==61.2.0
swissknife==0.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.11.3
utils==1.0.1
